Remove onClickSave
Adding code to change the style of the page will be performed
in a separate script file, since the highlighting of texts should
be done whenever a webpage is loaded, not just when the context
menu item is created. Content scripts run in the context of web
pages and can change them, which is exactly what we want.