Add functionality to remove saved texts
Users should be able to remove any selected text at will. With this commit, selected text that is alreay in storage will be removed from storage when the user clicks the context menu item 'highlight'. This does not currently support changing the style to reflect this change immediately -- the page must be reloaded.